updates to COLT
